feat: add dotenv dependency and implement WebSocket connection for scanner management
